[unrated] Title: 크기가 작은 부분문자열, Time: 0.13 ms, Memory: 72.2 MB -BaekjoonHub

+++ 프로그래머스/unrated/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.java
@@ -6,7 +6,7 @@ class Solution {
         List<String> s = new ArrayList<>();
         for (int i = 0; i < t.length() - (p.length() - 1); i++) {
             s.add(t.substring(i, i + p.length()));
-            if(Double.parseDouble(p) >= Double.parseDouble(s.get(i))) answer++;
+            if(Long.parseLong(p) >= Long.parseLong(s.get(i))) answer++;
         }
         return answer;
     }
